Separate jenkins integ test and unit test
